# api/openapi/tars-mvp.yaml
openapi: 3.1.0
info:
  title: TARS MVP API
  version: 0.1.0
  description: |
    Draft API definition aligned with tars_technical_design.md v1.4.
    Tag names describe API contract groupings, not Web Console page ownership.
    In particular, the `ops` tag means authenticated control-plane APIs and is
    not equivalent to the `/ops` UI page; some `ops`-tagged APIs are primarily
    surfaced from other daily object pages such as `/identity/*`, `/providers`,
    and `/connectors`.
servers:
  - url: https://tars.example.com
...
paths:
...
  /api/v1/roles:
    get:
      tags: [ops]
      summary: List roles
      operationId: listRoles
      security:
        - OpsBearer: []
      responses:
        "200":
          description: Role list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RoleListResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      tags: [ops]
      summary: Create role
      operationId: createRole
      security:
        - OpsBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RoleUpsertRequest"
      responses:
        "201":
          description: Role created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Role"
        "400":
          $ref: "#/components/responses/ValidationFailed"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/roles/{role_id}:
    get:
      tags: [ops]
      summary: Get role
      operationId: getRole
      security:
        - OpsBearer: []
      parameters:
        - name: role_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Role detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Role"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      tags: [ops]
      summary: Update role
      operationId: updateRole
      security:
        - OpsBearer: []
      parameters:
        - name: role_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RoleUpsertRequest"
      responses:
        "200":
          description: Updated role
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Role"
        "400":
          $ref: "#/components/responses/ValidationFailed"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /api/v1/agent-roles:
    get:
      tags: [ops]
      summary: List agent roles
      operationId: listAgentRoles
      parameters:
        - name: q
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
        - name: limit
          in: query
          schema:
            type: integer
        - name: sort_by
          in: query
          schema:
            type: string
        - name: sort_order
          in: query
          schema:
            type: string
            enum: [asc, desc]
      responses:
        "200":
          description: Paginated list of agent roles
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentRoleListResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
    post:
      tags: [ops]
      summary: Create an agent role
      operationId: createAgentRole
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AgentRole"
      responses:
        "201":
          description: Agent role created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentRole"
        "400":
          $ref: "#/components/responses/ValidationFailed"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          description: Role ID conflict
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /api/v1/agent-roles/{role_id}:
    get:
      tags: [ops]
      summary: Get an agent role by ID
      operationId: getAgentRole
      parameters:
        - name: role_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Agent role detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentRole"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      tags: [ops]
      summary: Update an agent role
      operationId: updateAgentRole
      parameters:
        - name: role_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AgentRole"
      responses:
        "200":
          description: Agent role updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentRole"
        "400":
          $ref: "#/components/responses/ValidationFailed"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [ops]
      summary: Delete an agent role
      operationId: deleteAgentRole
      parameters:
        - name: role_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Agent role deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/agent-roles/{role_id}/enable:
    post:
      tags: [ops]
      summary: Enable an agent role
      operationId: enableAgentRole
      parameters:
        - name: role_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Agent role enabled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentRole"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/agent-roles/{role_id}/disable:
    post:
      tags: [ops]
      summary: Disable an agent role
      operationId: disableAgentRole
      parameters:
        - name: role_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Agent role disabled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AgentRole"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
components:
  securitySchemes:
    OpsBearer:
      type: http
      scheme: bearer
      bearerFormat: opaque
  responses:
    Unauthorized:
      description: Authentication required or token invalid
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    Forbidden:
      description: Authenticated but not permitted
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
...
    ValidationFailed:
      description: Validation failed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
...
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: object
              additionalProperties: true
...
    Role:
      type: object
      properties:
        id:
          type: string
        display_name:
          type: string
        permissions:
          type: array
          items:
            type: string
    RoleListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Role"
...
    RoleUpsertRequest:
      type: object
      required: [role]
      properties:
        role:
          $ref: "#/components/schemas/Role"
        operator_reason:
          type: string
          description: Optional on IAM day-to-day role pages; backend may infer or default the audit reason.
...
    AgentRoleListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/AgentRole"
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        has_next:
          type: boolean
    AgentRole:
      type: object
      required:
        [
          role_id,
          display_name,
          status,
          is_builtin,
          profile,
          capability_binding,
          policy_binding,
        ]
      properties:
        role_id:
          type: string
        display_name:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [enabled, disabled]
        is_builtin:
          type: boolean
        profile:
          $ref: "#/components/schemas/AgentRoleProfile"
        capability_binding:
          $ref: "#/components/schemas/AgentRoleCapabilityBinding"
        policy_binding:
          $ref: "#/components/schemas/AgentRolePolicyBinding"
        model_binding:
          $ref: "#/components/schemas/AgentRoleModelBinding"
        org_id:
          type: string
        tenant_id:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    AgentRoleProfile:
      type: object
      properties:
        system_prompt:
          type: string
        persona_tags:
          type: array
          items:
            type: string
    AgentRoleCapabilityBinding:
      type: object
      required: [mode]
      properties:
        allowed_connector_capabilities:
          type: array
          items:
            type: string
        denied_connector_capabilities:
          type: array
          items:
            type: string
        allowed_skills:
          type: array
          items:
            type: string
        allowed_skill_tags:
          type: array
          items:
            type: string
        mode:
          type: string
          enum: [whitelist, blacklist, unrestricted]
    AgentRolePolicyBinding:
      type: object
      required: [max_risk_level, max_action]
      properties:
        max_risk_level:
          type: string
          enum: [info, warning, critical]
        max_action:
          type: string
          enum: [suggest_only, require_approval, direct_execute]
        require_approval_for:
          type: array
          items:
            type: string
        hard_deny:
          type: array
          items:
            type: string
    AgentRoleModelBinding:
      type: object
      properties:
        primary:
          $ref: "#/components/schemas/AgentRoleModelTargetBinding"
        fallback:
          $ref: "#/components/schemas/AgentRoleModelTargetBinding"
        inherit_platform_default:
          type: boolean
    AgentRoleModelTargetBinding:
      type: object
      properties:
        provider_id:
          type: string
        model:
          type: string